pre finish

+++ controller/auth.controller.js
@@ -10,8 +10,17 @@ const Complain = require('./../model/compain.model')
 
 exports.createAccount = async(req, res, next)=>{
     const {username, email, password, address, phone,key} = req.body;
+    const checkingAlreadyExist = await SignUp.find({email : email})
+    if(checkingAlreadyExist.length>0){
+        res.status(200).json({
+            status : 'error',
+            data : {
+                message : `This Account already exists`
+            }
+        }) 
+        return 
+    }
     const passwordCrypted = await bcrypt.hash(password, 12)
-
     const createAcc = await SignUp.create({username, email, password : passwordCrypted, address, phone, key})
     const token = await jwt.sign({id : createAcc._id},process.env.STRING)
     localStorage.setItem('b2cToken', token)
@@ -29,11 +38,21 @@ exports.createAccount = async(req, res, next)=>{
 exports.Login = async(req, res, next)=>{
     const {email, password} = req.body;
     const findingUser = await SignUp.find({email : email})
+    if(findingUser.length===0){
+        res.status(200).json({
+            status : 'error',
+            data : {
+                message : 'invalid email or password'
+            }
+        })
+        return;
+    }    
     // checking password
     const passCheck = await bcrypt.compare(password,findingUser[0].password)
+    
     if(!passCheck){
         res.status(200).json({
-            status : 'success',
+            status : 'error',
             data : {
                 
                 message : 'invalid email or password'
@@ -56,6 +75,15 @@ exports.Login = async(req, res, next)=>{
 exports.forgotPassword = async(req, res, next)=>{
     const {email,key,password} = req.body;
     const findingUser = await SignUp.find({email : email})
+    if(findingUser.length===0){
+        res.status(200).json({
+            status : 'error',
+            data : {
+                message : 'invalid email'
+            }
+        })
+        return;
+    } 
     if(findingUser[0].key === key){
         findingUser[0].password = await bcrypt.hash(password, 12)
         const token = await jwt.sign({id : findingUser[0]._id},process.env.STRING)
@@ -91,6 +119,17 @@ exports.LoggingOut = async(req, res,next)=>{
 exports.cartPreperation = async(req, res, next)=>{
     const {itemName, units, price} = req.body;
     const token =localStorage.getItem('b2cToken')
+    if(!token){
+        res.status(200).json({
+            status : 'success',
+            data : {
+                cart : [],
+                nameArr : [],
+                message : 'please login or create your account first'
+            }
+        }) 
+        return 
+    }
     const decode = await promisify(jwt.verify)(token, process.env.STRING)
     const findingUser = await SignUp.findById(decode.id)
     
@@ -106,13 +145,35 @@ exports.cartPreperation = async(req, res, next)=>{
         nameArr.push(el.itemName)
     })
 
-    
+    const itemsSendableArray = []
+    const items = await Product.find()
+    items.forEach(el=>{
+        const obj = {
+            name : el.name,
+            price : el.price,
+            quantity : 0,
+            variety : el.variety
+        }
+        itemsSendableArray.push(obj)
+    })
+    const cart = findingUser.cart
+    let pricegross = 0;
+    cart.forEach(el=>{
+        itemsSendableArray.forEach(item=>{
+            if(item.name===el.itemName){
+                item.quantity = el.units
+            }
+        })
+        pricegross = Number(el.price.split('/-')[0]*el.units) + pricegross
+    })
 
     res.status(200).json({
         status : 'success',
         data : {
+            items : itemsSendableArray,
             cart : findingUser.cart,
             nameArr,
+            pricegross,
             message :  `${itemName} added , you can change quantity in cart section`
         }
     })
@@ -121,6 +182,18 @@ exports.cartPreperation = async(req, res, next)=>{
 
 exports.sendingAlreadyAddedCart = async(req, res, next)=>{
     const token =localStorage.getItem('b2cToken')
+    if(!token){
+        res.status(200).json({
+            status : 'success',
+            data : {
+                cart :[],
+                combo: [],
+                subTotal : 0,
+                total : 0
+            }
+        })  
+        return
+    }
     const decode = await promisify(jwt.verify)(token, process.env.STRING)
     const findingUser = await SignUp.findById(decode.id)
     let cart = findingUser.cart
@@ -132,10 +205,22 @@ exports.sendingAlreadyAddedCart = async(req, res, next)=>{
     const tax = 0;
     const platform = 2;
     const total = grossprice + delivery + grossprice*(tax/100) + platform
+
+    // combos
+    const combo = findingUser.combos;
+    const allComboitemList = combo.map(el=>{
+        return el.list;
+    })
+    const distributinglist = []
+    allComboitemList.forEach(el=>{
+        distributinglist.push(...el)
+    })
+
     res.status(200).json({
         status : 'success',
         data : {
             cart :cart,
+            combo,
             subTotal : grossprice,
             total
         }
@@ -147,6 +232,15 @@ exports.updatingUnits = async(req, res, next)=>{
     const {units, name} = req.body;
     console.log(units)
     const token =localStorage.getItem('b2cToken')
+    if(!token){
+        res.status(200).json({
+            status : 'success',
+            data : {
+                message : 'please login or open your account first'
+            }
+        })  
+        return
+    }
     const decode = await promisify(jwt.verify)(token, process.env.STRING)
     const findingUser = await SignUp.findById(decode.id)
     const cart = findingUser.cart
@@ -157,10 +251,35 @@ exports.updatingUnits = async(req, res, next)=>{
     })
 
     findingUser.save()
+
+    const itemsSendableArray = []
+    const items = await Product.find()
+    items.forEach(el=>{
+        const obj = {
+            name : el.name,
+            price : el.price,
+            quantity : 0,
+            variety : el.variety
+        }
+        itemsSendableArray.push(obj)
+    })
+    let price = 0
+    cart.forEach(el=>{
+        itemsSendableArray.forEach(item=>{
+            if(item.name===el.itemName){
+                item.quantity = el.units
+            }
+        })
+        price = Number(el.price.split('/-')[0]*el.units) + price
+    })
+    console.log(price)
+
     res.status(200).json({
         status : 'success',
         data : {
-            message : 'updated'
+            items : itemsSendableArray,
+            message : 'updated',
+            price
         }
     })
 }
@@ -187,6 +306,34 @@ exports.deletingCartItem = async(req, res, next)=>{
     })
 }
 
+exports.deleteCompleteCart = async(req, res, next)=>{
+    const token =localStorage.getItem('b2cToken')
+    const decode = await promisify(jwt.verify)(token, process.env.STRING)
+    const findingUser = await SignUp.findById(decode.id)
+    findingUser.cart = []
+    findingUser.save()
+    res.status(200).json({
+        status : 'success',
+        data : {
+            cart : findingUser.cart,
+            message : 'your cart is empty now!' 
+        }
+    })
+}
+
+exports.sendingCartItemsOnlyWithoutCalculations = async(req, res, next)=>{
+    const token =localStorage.getItem('b2cToken')
+    const decode = await promisify(jwt.verify)(token, process.env.STRING)
+    const findingUser = await SignUp.findById(decode.id)
+    const cart = findingUser.cart
+    res.status(200).json({
+        status : 'success',
+        data : {
+            cart
+        }
+    })
+}
+
 exports.onlinePayment = async(req, res, next)=>{
     const token =localStorage.getItem('b2cToken')
     const decode = await promisify(jwt.verify)(token, process.env.STRING)
@@ -207,6 +354,7 @@ exports.onlinePayment = async(req, res, next)=>{
         currency: "INR",
         receipt: "order_rcptid_11"
     };
+
     instance.orders.create(options, function(err, order) {
         console.log(order);
         res.status(200).json({
@@ -219,49 +367,66 @@ exports.onlinePayment = async(req, res, next)=>{
 
 }
 
+const crypto = require('crypto')
 
 exports.validatingPays = async(req, res, next)=>{
     const {razorpay_payment_id, razorpay_order_id, razorpay_signature} = req.body;
-    const token =localStorage.getItem('b2cToken')
-    const decode = await promisify(jwt.verify)(token, process.env.STRING)
-    const findingUser = await SignUp.findById(decode.id)
-    const cart = findingUser.cart
-    let grossprice = 0
-    cart.forEach(el=>{
-        grossprice = Number(grossprice) + Number(el.price.split('/-')[0]*el.units)
-    })
-    const delivery = 20;
-    const tax = 0;
-    const platform = 2;
-    const total = grossprice + delivery + grossprice*(tax/100) + platform
     
-    findingUser.placed_orders.push({
-        item_list : findingUser.cart,
-        subtotal : grossprice,
-        delivery : delivery,
-        tax : tax,
-        platform : platform,
-        total : total,
-        razorpay_payment_id,
-        razorpay_order_id,
-        razorpay_signature,
-        payment_mode : 'online',
-        payment_status : 'success',
-        creds : [findingUser.username, findingUser.address, findingUser.phone],
-        client_id : findingUser._id
-    })
+    var instance = new Razorpay({ key_id: process.env.RAZORPAY_ID, key_secret: process.env.RAZORPAY_SECRET })
+    let hmac = crypto.createHmac('sha256', process.env.RAZORPAY_SECRET); 
+    hmac.update(razorpay_order_id + "|" + razorpay_payment_id)
+    const generated_signature = hmac.digest('hex')
+    if(razorpay_signature===generated_signature){ 
+        const token =localStorage.getItem('b2cToken')
+        const decode = await promisify(jwt.verify)(token, process.env.STRING)
+        const findingUser = await SignUp.findById(decode.id)
+        const cart = findingUser.cart
+        let grossprice = 0
+        cart.forEach(el=>{
+            grossprice = Number(grossprice) + Number(el.price.split('/-')[0]*el.units)
+        })
+        const delivery = 20;
+        const tax = 0;
+        const platform = 2;
+        const total = grossprice + delivery + grossprice*(tax/100) + platform
+        
+        findingUser.placed_orders.push({
+            item_list : findingUser.cart,
+            subtotal : grossprice,
+            delivery : delivery,
+            tax : tax,
+            platform : platform,
+            total : total,
+            razorpay_payment_id,
+            razorpay_order_id,
+            razorpay_signature,
+            payment_mode : 'online',
+            payment_status : 'success',
+            creds : [findingUser.username, findingUser.address, findingUser.phone],
+            client_id : findingUser._id
+        })
+    
+        
+        findingUser.cart = []
+        
+        findingUser.save()
+    
+        res.status(200).json({
+            status : 'success',
+            data : {
+                message : 'order placed sucessfully'
+            }
+        })
 
-    
-    findingUser.cart = []
-    
-    findingUser.save()
+    } else{
+        res.status(200).json({
+            status : 'success',
+            data : {
+                message : 'Payment failed please try again or place cash on delivery order'
+            }
+        })
+    }
 
-    res.status(200).json({
-        status : 'success',
-        data : {
-            message : 'order placed sucessfully'
-        }
-    })
 }
 
 
@@ -309,13 +474,25 @@ exports.codAndPlacingOrder = async(req, res, next)=>{
 
 exports.sendingPlacedOrder = async(req, res, next)=>{
     const token =localStorage.getItem('b2cToken')
+    if(!token){
+        res.status(200).json({
+            status : 'success',
+            data : {
+                placed_orders : [],
+                message : 'Please login to see all placed orders'
+            }
+        })  
+        return
+    }
     const decode = await promisify(jwt.verify)(token, process.env.STRING)
     const findingUser = await SignUp.findById(decode.id)
     const placed_orders = findingUser.placed_orders
     res.status(200).json({
         status : 'success',
         data : {
-            placed_orders
+            placed_orders,
+            message : "You don't have any placed orders"
+
         }
     })
 
@@ -325,6 +502,16 @@ exports.sendingPlacedOrder = async(req, res, next)=>{
 exports.feedback = async(req, res, next)=>{
     const {feedbackMessage} = req.body;
     const token =localStorage.getItem('b2cToken')
+    if(!token){
+        res.status(200).json({
+            status : 'success',
+            data : {
+                message : 'Please login or create account first'
+
+            }
+        })  
+        return
+    }
     const decode = await promisify(jwt.verify)(token, process.env.STRING)
     const findingUser = await SignUp.findById(decode.id)
     findingUser.feedback.push({
@@ -345,6 +532,16 @@ exports.feedback = async(req, res, next)=>{
 exports.complain = async(req, res, next)=>{
     const {subject, description} = req.body;
     const token =localStorage.getItem('b2cToken')
+    if(!token){
+        res.status(200).json({
+            status : 'success',
+            data : {
+                message : 'Please login or create account first'
+
+            }
+        })  
+        return
+    }
     const decode = await promisify(jwt.verify)(token, process.env.STRING)
     const findingUser = await SignUp.findById(decode.id)
     const complaincreate = await Complain.create({subject, description, user : findingUser._id})
@@ -354,4 +551,92 @@ exports.complain = async(req, res, next)=>{
             message : 'Complain registered our team will contact you in next 24hrs'
         }
     })
+}
+
+
+exports.PuttingComboToCart = async(req, res, next)=>{
+    const {list} = req.body;
+    const token =localStorage.getItem('b2cToken')
+    if(!token){
+        res.status(200).json({
+            status : 'success',
+            data : {
+                message : 'Please login or create account first'
+
+            }
+        })  
+        return
+    }
+    const decode = await promisify(jwt.verify)(token, process.env.STRING)
+    const findingUser = await SignUp.findById(decode.id)
+    findingUser.combos.push({list})
+    console.log(list)
+    findingUser.save()
+    res.status(200).json({
+        status : 'success',
+        data : {
+            message : 'combo added'
+        }
+    })
+}
+
+exports.sendingUSerData = async(req, res, next)=>{
+    const token =localStorage.getItem('b2cToken')
+    if(!token){
+        res.status(200).json({
+            status : 'success',
+            data : {
+                user : {},
+                message : 'Please login or create account first'
+
+            }
+        })  
+        return
+    }
+    const decode = await promisify(jwt.verify)(token, process.env.STRING)
+    const findingUser = await SignUp.findById(decode.id)
+    res.status(200).json({
+        status : 'success',
+        data : {
+            user : findingUser
+        }
+    })
+}
+
+exports.changeAddress = async(req, res, next)=>{
+    const {address} = req.body;
+    const token =localStorage.getItem('b2cToken')
+    const decode = await promisify(jwt.verify)(token, process.env.STRING)
+    const findingUser = await SignUp.findById(decode.id)
+    findingUser.address = address
+    res.status(200).json({
+        status : 'success',
+        data : {
+            user : findingUser
+        }
+    })
+}
+
+var timeout = require('connect-timeout')
+
+exports.code = async(req, res, next)=>{
+
+    const accountSid = 'AC7526bd4af40d83f63312572f54b61c58';
+const authToken = '92d3f8e4b66b1930d83cfe693824c573';
+const client = require('twilio')(accountSid, authToken);
+const otp = Math.trunc(Math.random()*778884)
+client.messages
+    .create({
+        body: `Welcome to your OTP is ${otp}`,
+        from: 'whatsapp:+14155238886',
+        to: 'whatsapp:+919914153088'
+    })
+    .then(message => console.log(message.sid))
+    
+    
+    
+    res.status(200).json({
+        status : 'success'
+    })
+
 }
